+++ src/main/java/org/scoooting/dto/MotorcycleDTO.java
@@ -6,5 +6,5 @@ import org.scoooting.entities.enums.MotorcycleStatus;
 import java.math.BigDecimal;
 
 @Builder
-public record MotorcycleDTO(Long id, String model, MotorcycleStatus status, Double latitude, Double longitude,
+public record MotorcycleDTO(Long id, String model, MotorcycleStatus status, Float latitude, Float longitude,
                             Integer engineSize, BigDecimal fuelLevel) {}
