+++ src/main/java/org/scoooting/dto/ScootersDTO.java
@@ -2,4 +2,4 @@ package org.scoooting.dto;
 
 import org.scoooting.entities.enums.ScootersStatus;
 
-public record ScootersDTO( Long id, String model, ScootersStatus status, Double latitude, Double longitude) {}
+public record ScootersDTO( Long id, String model, ScootersStatus status, Float latitude, Float longitude) {}
